build(auditlog): auditlog refactoring
