+++ libs/shared/src/components/auditlog/AuditLogFilter.tsx
@@ -1,0 +1,374 @@
+import { ChevronDownIcon } from "lucide-react";
+import { useCallback, useEffect, useMemo, useState } from "react";
+import { runPixel } from "@semoss/sdk";
+import {
+	Button,
+	Calendar,
+	type DateRange,
+	DropdownMenu,
+	DropdownMenuCheckboxItem,
+	DropdownMenuContent,
+	DropdownMenuRadioGroup,
+	DropdownMenuSeparator,
+	DropdownMenuTrigger,
+	Input,
+	Popover,
+	PopoverAnchor,
+	PopoverContent,
+} from "@semoss/ui/next";
+import { dateFormat, ENGINE_TYPES } from "./common";
+
+const initialAcc = {
+	APP: [],
+	MODEL: [],
+	DATABASE: [],
+	VECTOR: [],
+	FUNCTION: [],
+	STORAGE: [],
+};
+
+const DashboardDurations = [
+	{ label: "Today", value: "today", dateRangeType: "DAY", dateRangeValue: 1 },
+	{
+		label: "Last 7 Days",
+		value: "last7days",
+		dateRangeType: "WEEK",
+		dateRangeValue: 1,
+	},
+	{
+		label: "Last 30 Days",
+		value: "last30days",
+		dateRangeType: "MONTH",
+		dateRangeValue: 1,
+	},
+	{
+		label: "Custom",
+		value: "custom",
+		renderWithSeparator: true,
+		dateRangeType: "CUSTOM",
+		dateRangeValue: 1,
+	},
+];
+
+export const AuditLogFilter = (props) => {
+	const { insightId, updateLogs } = props;
+	const [engineDetails, setEngineDetails] = useState({ ...initialAcc });
+	const [engineSelectionDetails, setEngineSelectionDetails] = useState({
+		engineType: "",
+		engineId: "",
+	});
+	const [dashboardDuration, setDashboardDuration] =
+		useState<(typeof DashboardDurations)[number]["value"]>("");
+	const [showCustomPopover, setShowCustomPopover] = useState<boolean>(false);
+	const [customDateRange, setCustomDateRange] = useState<DateRange | null>({
+		from: new Date(),
+		to: new Date(),
+	});
+	useEffect(() => {
+		async function getMyEngines() {
+			if (insightId) {
+				const response = await runPixel(`MyEngines();`, insightId);
+				const responseData = response.pixelReturn[0].output;
+				let enginesDropdown = (
+					responseData as Array<{
+						database_id: string;
+						app_type: string;
+						app_name: string;
+					}>
+				).reduce(
+					(acc, engine) => {
+						// Only accept known app_types
+						if (Object.hasOwn(acc, engine.app_type)) {
+							acc[engine.app_type] = [
+								...acc[engine.app_type],
+								{
+									value: engine.database_id,
+									label: engine.app_name,
+								},
+							];
+						}
+						return acc;
+					},
+					{ ...initialAcc },
+				);
+				const projectResponse = await runPixel(
+					`MyProjects();`,
+					insightId,
+				);
+				const projectResponseData =
+					projectResponse.pixelReturn[0].output;
+				const projectsDropdown = (
+					projectResponseData as Array<{
+						project_id: string;
+						project_type: string;
+						project_name: string;
+					}>
+				).reduce(
+					(acc, engine) => {
+						// Only accept known app_types
+						if (Object.hasOwn(acc, "APP")) {
+							acc.APP = [
+								...acc.APP,
+								{
+									value: engine.project_id,
+									label: engine.project_name,
+								},
+							];
+						}
+						return acc;
+					},
+					{ ...initialAcc },
+				);
+				enginesDropdown = {
+					...enginesDropdown,
+					APP: projectsDropdown.APP,
+				};
+				setEngineDetails((prev) => ({ ...prev, ...enginesDropdown }));
+			}
+		}
+		getMyEngines();
+	}, [insightId]);
+
+	const SelectedDuration = useMemo(() => {
+		return (
+			DashboardDurations.find(
+				(duration) => duration.value === dashboardDuration,
+			) || { label: "", value: "", dateRangeType: "", dateRangeValue: 1 }
+		);
+	}, [dashboardDuration]);
+
+	useEffect(() => {
+		// Implementation for triggering logs API
+		if (
+			engineSelectionDetails.engineId !== "" &&
+			dashboardDuration !== ""
+		) {
+			updateLogs({
+				engineId: engineSelectionDetails.engineId,
+				duration: dashboardDuration,
+				customDateRange: customDateRange,
+				engineType: engineSelectionDetails.engineType,
+				SelectedDuration: SelectedDuration,
+			});
+		}
+	}, [
+		engineSelectionDetails.engineId,
+		dashboardDuration,
+		customDateRange,
+		engineSelectionDetails.engineType,
+		SelectedDuration,
+	]);
+
+	const renderCustomDatePopover = useCallback(() => {
+		if (!showCustomPopover) return null;
+		return (
+			<Popover open={showCustomPopover}>
+				<PopoverAnchor>
+					<PopoverContent className="flex w-[75%] flex-col gap-4 p-4">
+						<div className="flex justify-between gap-2">
+							<Input
+								value={dateFormat(
+									customDateRange.from?.toString(),
+								)}
+								type="text"
+								className="w-[50%]"
+							></Input>
+							<Input
+								value={dateFormat(
+									customDateRange.to?.toString(),
+								)}
+								type="text"
+								className="w-[50%]"
+							></Input>
+						</div>
+						<div className="flex justify-around">
+							<Calendar
+								mode="range"
+								selected={customDateRange}
+								onSelect={(daterange) => {
+									if (daterange?.from && daterange?.to) {
+										setCustomDateRange(daterange);
+									}
+								}}
+								className="rounded-md border shadow-sm"
+								captionLayout="dropdown"
+								timeZone="UTC"
+							/>
+						</div>
+						<div className="flex justify-end">
+							<Button
+								variant="outline"
+								className="w-fit justify-end bg-primary text-white"
+								size="sm"
+								onClick={() => {
+									setShowCustomPopover(false);
+									if (
+										customDateRange?.from &&
+										customDateRange?.to &&
+										engineSelectionDetails.engineId
+									) {
+										updateLogs();
+									}
+								}}
+							>
+								Apply
+							</Button>
+						</div>
+					</PopoverContent>
+				</PopoverAnchor>
+			</Popover>
+		);
+	}, [showCustomPopover, customDateRange, engineSelectionDetails.engineId]);
+
+	return (
+		<div className="flex gap-2">
+			<div className="flex min-w-[100px] justify-between">
+				<DropdownMenu>
+					<DropdownMenuTrigger
+						asChild
+						className="flex justify-between align-center"
+					>
+						<Button
+							variant="outline"
+							size="sm"
+							className={`flex min-w-[180px] justify-between`}
+						>
+							<div className="flex w-full justify-between align-center">
+								<span className="flex justify-start">
+									{engineSelectionDetails.engineType !== ""
+										? engineSelectionDetails.engineType
+										: "Select Engine Type"}{" "}
+								</span>
+								<ChevronDownIcon className="flex justify-end align-center" />
+							</div>
+						</Button>
+					</DropdownMenuTrigger>
+					<DropdownMenuContent>
+						<DropdownMenuRadioGroup>
+							{ENGINE_TYPES.map((engineType) => (
+								<DropdownMenuCheckboxItem
+									checked={
+										engineType ===
+										engineSelectionDetails.engineType
+									}
+									key={`${engineType}Selection`}
+									onCheckedChange={(open) => {
+										setEngineSelectionDetails({
+											...engineSelectionDetails,
+											engineType: open ? engineType : "",
+											engineId: "",
+										});
+									}}
+								>
+									{engineType}
+								</DropdownMenuCheckboxItem>
+							))}
+						</DropdownMenuRadioGroup>
+					</DropdownMenuContent>
+				</DropdownMenu>
+			</div>
+			<div className="flex min-w-[100px] justify-between">
+				<DropdownMenu>
+					<DropdownMenuTrigger
+						asChild
+						className="flex justify-between align-center"
+					>
+						<Button
+							variant="outline"
+							size="sm"
+							className={`flex min-w-[180px] justify-between`}
+						>
+							<div className="flex w-full justify-between">
+								<span className="flex justify-start">
+									{engineDetails?.[
+										engineSelectionDetails?.engineType
+									]
+										?.filter(
+											(engine) =>
+												engine.value ===
+												engineSelectionDetails.engineId,
+										)
+										.map((engine) => engine.label) ??
+										"Select Engine"}{" "}
+								</span>
+								<ChevronDownIcon className="flex justify-end align-center" />
+							</div>
+						</Button>
+					</DropdownMenuTrigger>
+					<DropdownMenuContent>
+						<DropdownMenuRadioGroup>
+							{engineSelectionDetails.engineType &&
+								engineDetails[engineSelectionDetails.engineType]
+									.length > 0 &&
+								engineDetails[
+									engineSelectionDetails.engineType
+								].map((engine) => (
+									<DropdownMenuCheckboxItem
+										key={engine.value}
+										checked={
+											engine.value ===
+											engineSelectionDetails.engineId
+										}
+										onCheckedChange={(prop) => {
+											setEngineSelectionDetails({
+												...engineSelectionDetails,
+												engineId: prop
+													? engine.value
+													: "",
+											});
+										}}
+									>
+										{engine.label}
+									</DropdownMenuCheckboxItem>
+								))}
+						</DropdownMenuRadioGroup>
+					</DropdownMenuContent>
+				</DropdownMenu>
+			</div>
+			<div className="min-w-[100px]">
+				<DropdownMenu>
+					<DropdownMenuTrigger asChild>
+						<Button variant="outline" size="sm">
+							{SelectedDuration?.label === ""
+								? "Today"
+								: SelectedDuration?.label}{" "}
+							<ChevronDownIcon />
+						</Button>
+					</DropdownMenuTrigger>
+					<DropdownMenuContent>
+						{/* Add dropdown items here */}
+						<DropdownMenuRadioGroup>
+							{DashboardDurations.map((duration) => (
+								<>
+									{duration.renderWithSeparator && (
+										<DropdownMenuSeparator />
+									)}
+									<DropdownMenuCheckboxItem
+										key={duration.value}
+										checked={
+											duration.value === dashboardDuration
+										}
+										onCheckedChange={() => {
+											setDashboardDuration(
+												duration.value,
+											);
+											if (duration.value === "custom") {
+												setShowCustomPopover(true);
+											} else {
+												if (showCustomPopover)
+													setShowCustomPopover(false);
+											}
+										}}
+									>
+										{duration.label}
+									</DropdownMenuCheckboxItem>
+								</>
+							))}
+						</DropdownMenuRadioGroup>
+					</DropdownMenuContent>
+				</DropdownMenu>
+			</div>
+			{renderCustomDatePopover()}
+		</div>
+	);
+};
