+++ libs/shared/src/components/auditlog/common.ts
@@ -52,3 +52,21 @@ export const TimeDateFormatter = (
 		return { date: "", time: "" };
 	}
 };
+
+export const dateFormat = (dateString: string | undefined) => {
+	if (!dateString) return "";
+	const date = new Date(dateString);
+	const year = date.getFullYear();
+	const month = String(date.getMonth() + 1).padStart(2, "0");
+	const day = String(date.getDate()).padStart(2, "0");
+	return `${year}-${month}-${day}`;
+};
+
+export const ENGINE_TYPES = [
+	"APP",
+	"MODEL",
+	"DATABASE",
+	"VECTOR",
+	"FUNCTION",
+	"STORAGE",
+];
