+++ libs/shared/src/components/auditlog/index.ts
@@ -1,4 +1,6 @@
+import { AuditLogFilter } from "./AuditLogFilter";
 import { AuditLogsDataTable } from "./AuditLogsDataTable";
 import { AuditLogsTimeline } from "./AuditLogsTimeline";
+import { dateFormat } from "./common";
 
-export { AuditLogsDataTable, AuditLogsTimeline };
+export { AuditLogsDataTable, AuditLogsTimeline, AuditLogFilter, dateFormat };
