+++ packages/auditlog/src/components/common/utility.ts
@@ -52,12 +52,3 @@ export const TimeDateFormatter = (
 		return { date: "", time: "" };
 	}
 };
-
-export const ENGINE_TYPES = [
-	"APP",
-	"MODEL",
-	"DATABASE",
-	"VECTOR",
-	"FUNCTION",
-	"STORAGE",
-];
